services/AgendaAnalytics-Mapping/src/requirements.txt
anyio==3.6.1
asyncpg==0.20.1
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.5
databases==0.2.6
dynaconf==3.1.9
fastapi>=0.78.0
fastapi-utils>=0.2.1
fastapi-mqtt>=1.0.6
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.10
iniconfig==1.1.1
jmespath==1.0.1
numpy==1.23.0
packaging==21.3
pandas==1.4.2
paho-mqtt>=1.6.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
requests-mock==1.9.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn>=0.11.2
uuid==1.30
websockets>=9.0
prometheus-fastapi-instrumentator
opentelemetry-instrumentation-fastapi==0.29b0
opentelemetry-instrumentation-logging==0.29b0
opentelemetry-exporter-otlp==1.10.0
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
protobuf==3.20.*